backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-autofill-backend"
version = "1.0.0"
description = "CV Auto-Fill Backend API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "hiredis==2.3.2",
    "openai==1.10.0",
    "anthropic==0.18.1",
    "httpx==0.26.0",
    "pypdf2==3.0.1",
    "pdfplumber==0.10.3",
    "pdf2image==1.16.3",
    "Pillow==10.2.0",
    "python-dotenv==1.0.1",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    "fitz==0.0.1.dev2",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.26.0",
]

[tool.setuptools]
packages = ["app"]
